+++ app/src/main/java/com/keystone/cold/viewmodel/EthTxConfirmViewModel.java
@@ -34,10 +34,9 @@ import com.keystone.coinlib.exception.InvalidPathException;
 import com.keystone.coinlib.exception.InvalidTransactionException;
 import com.keystone.coinlib.interfaces.SignCallback;
 import com.keystone.coinlib.interfaces.Signer;
-import com.keystone.coinlib.model.Contract;
 import com.keystone.coinlib.path.CoinPath;
-import com.keystone.coinlib.utils.ContactExternalDbLoader;
 import com.keystone.coinlib.utils.Coins;
+import com.keystone.coinlib.utils.ContactExternalDbLoader;
 import com.keystone.cold.AppExecutors;
 import com.keystone.cold.R;
 import com.keystone.cold.callables.ClearTokenCallable;
@@ -116,7 +115,7 @@ public class EthTxConfirmViewModel extends TxConfirmViewModel {
                 if (!TextUtils.isEmpty(abiFile)) {
                     addressSymbol = abiFile.replace(".json", "");
                 } else {
-                    Contract dataFromTFCard = ContactExternalDbLoader.contractData(to);
+                    ContactExternalDbLoader.Contract dataFromTFCard = ContactExternalDbLoader.contractData(to);
                     addressSymbol = dataFromTFCard.getName();
                 }
             }
