+++ coinlib/src/main/java/com/keystone/coinlib/coins/ETH/EthImpl.java
@@ -28,9 +28,8 @@ import com.keystone.coinlib.coins.SignTxResult;
 import com.keystone.coinlib.interfaces.Coin;
 import com.keystone.coinlib.interfaces.SignCallback;
 import com.keystone.coinlib.interfaces.Signer;
-import com.keystone.coinlib.model.Contract;
-import com.keystone.coinlib.utils.ContactExternalDbLoader;
 import com.keystone.coinlib.utils.Coins;
+import com.keystone.coinlib.utils.ContactExternalDbLoader;
 
 import org.bouncycastle.util.encoders.Hex;
 import org.json.JSONException;
@@ -131,8 +130,8 @@ public class EthImpl implements Coin {
                 abi = readAsset("abi/" + abiFile);
                 contractName = abiFile.replace(".json", "");
             } else {
-                Contract dataFromTFCard = ContactExternalDbLoader.contractData(rawTx.getTo());
-                abi = getAbiFromContract(dataFromTFCard);
+                ContactExternalDbLoader.Contract dataFromTFCard = ContactExternalDbLoader.contractData(rawTx.getTo());
+                abi = dataFromTFCard.getAbi();
                 contractName = dataFromTFCard.getName();
                 if (!TextUtils.isEmpty(abi) && callback != null) {
                     callback.fromTFCard();
@@ -172,18 +171,6 @@ public class EthImpl implements Coin {
             return null;
         }
         return metaData;
-    }
-
-    private static String getAbiFromContract(Contract contract) {
-        String abi = null;
-        try {
-            JSONObject metaData = new JSONObject(contract.getMetadata());
-            JSONObject output = metaData.getJSONObject("output");
-            abi = output.getString("abi");
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-        return abi;
     }
 
     public static String getSignature(String signedHex) {
