--- coinlib/src/main/java/com/keystone/coinlib/model/Contract.java
@@ -1,23 +1,0 @@
-package com.keystone.coinlib.model;
-
-public class Contract {
-    private String name;
-
-    private String metadata;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getMetadata() {
-        return metadata;
-    }
-
-    public void setMetadata(String metadata) {
-        this.metadata = metadata;
-    }
-}
